BLOCKS-048 replaced tab with double space in blocks, adjusted test scripts to work with new folder structure [AK]
