+++ package.json
@@ -27,7 +27,9 @@
   "devDependencies": {
     "@babel/core": "^7.6.4",
     "@babel/preset-env": "^7.6.3",
+    "babel-core": "^6.26.3",
     "babel-loader": "^8.0.6",
+    "babel-preset-env": "^1.7.0",
     "copy-webpack-plugin": "^5.0.4",
     "core-js": "^3.3.4",
     "css-loader": "^3.2.0",
@@ -35,8 +37,11 @@
     "eslint-loader": "^3.0.2",
     "html-webpack-plugin": "^3.2.0",
     "jest": "^24.9.0",
+    "jest-environment-node": "^24.9.0",
     "jest-puppeteer": "^4.3.0",
     "mini-css-extract-plugin": "^0.8.0",
+    "puppeteer": "^2.0.0",
+    "rimraf": "^3.0.0",
     "webpack": "^4.41.2",
     "webpack-cli": "^3.3.9",
     "webpack-dev-server": "^3.9.0",
